se genera metodo de informacion de sucursales por usuario

+++ Back/ATV-Backend/office-service/axity.office-commons/src/main/java/com/axity/office/commons/dto/ProductDocumentDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ProductDocumentDTO {
-    private Long xpid;
+    private Integer xpid;
     private String xpcopr;
     private String xpcodo;
     private String xpdsdo;
